feat: Add kiro-acp-proxy, update nvim config, refactor sync.sh
Add kiro-acp-proxy script to ~/bin and update nvim config.

Refactor sync.sh:
- Add OS detection (macos/apt/yum) replacing pkg_install wrappers
- Extract focused functions: sync_tmux, sync_ripgrep, sync_tree_sitter,
  sync_nvim_config, sync_alacritty, sync_themes, sync_fonts,
  sync_apple_keyboard
- Add confirmation prompts to all sync functions
- Install alacritty from apt instead of building from source
- Support font installation on macOS via ~/Library/Fonts
- Skip Linux-only cargo installs (ripgrep, tree-sitter) on macOS
- Fix command -v bug in sync_cmake
- Remove dead commented-out code in sync_node

+++ sync.sh
@@ -11,20 +11,23 @@ export PATH="$HOME/.cargo/bin:$PATH"
 export PATH="$HOME/.local/bin:$PATH"
 export LD_LIBRARY_PATH="$HOME/.local/lib64:$LD_LIBRARY_PATH"
 
+# Detect OS
+if [ "$(uname)" = "Darwin" ]; then
+    os="macos"
+elif [ -x "$(command -v "apt")" ]; then
+    os="apt"
+elif [ -x "$(command -v "yum")" ]; then
+    os="yum"
+else
+    os="unknown"
+fi
+
 require_command() {
     if ! [ -x "$(command -v "$1")" ]; then
         echo "Couldn't find $1 on your system. I cannot continue..."
         exit 1
     fi
 }
-
-if [ -x "$(command -v "apt")" ]; then
-    os_pkg_manager="apt"
-fi
-
-if [ -x "$(command -v "yum")" ]; then
-    os_pkg_manager="yum"
-fi
 
 confirm() {
     echo "$1"
@@ -45,34 +48,12 @@ confirm() {
     esac
 }
 
-apt_install() {
-    if [ -x "$(command -v "$2")" ]; then
-        return
-    fi
-    if [ -x "$(command -v "$1")" ]; then
-        return
-    fi
-    if [ "$os_pkg_manager" != "apt" ]; then
-        return
-    fi
-    echo "Couldn't find $1. I'll try to install it..."
-    sudo apt -y install "$1"
-}
-
-yum_install() {
-    if [ -x "$(command -v "$1")" ]; then
-        return
-    fi
-    if [ "$os_pkg_manager" != "yum" ]; then
-        return
-    fi
-    echo "Couldn't find $1. I'll try to install it..."
-    sudo yum -y install "$1"
-}
-
 sync_git() {
-    apt_install "git"
-    yum_install "git"
+    if [ "$os" = "apt" ]; then
+        sudo apt -y install git
+    elif [ "$os" = "yum" ]; then
+        sudo yum -y install git
+    fi
     require_command "git"
 
     if confirm "I will now replace your git configuration."; then
@@ -82,8 +63,11 @@ sync_git() {
 }
 
 sync_shell() {
-    apt_install "zsh"
-    yum_install "zsh"
+    if [ "$os" = "apt" ]; then
+        sudo apt -y install zsh
+    elif [ "$os" = "yum" ]; then
+        sudo yum -y install zsh
+    fi
     require_command "zsh"
 
     if confirm "I will now reinstall oh-my-zsh and replace your zsh configuration."; then
@@ -100,44 +84,53 @@ sync_shell() {
             echo "You'll probably have to provide a password :("
             chsh -s "$(command -v zsh)"
         fi
+    fi
+}
 
-        apt_install "tmux"
-        yum_install "tmux"
+sync_tmux() {
+    if confirm "I will now replace your tmux configuration."; then
+        if [ "$os" = "apt" ]; then
+            sudo apt -y install tmux
+        elif [ "$os" = "yum" ]; then
+            sudo yum -y install tmux
+        fi
         rm -f "$HOME/.tmux.conf"
         curl -sSLo "$HOME/.tmux.conf" "$dotfiles_url/home/.tmux.conf"
     fi
 }
 
 sync_gcc() {
-    if [ "$os_pkg_manager" = "apt" ]; then
+    if [ "$os" != "yum" ]; then
         return
     fi
     if [ -x "$HOME/.local/bin/gcc" ]; then
         return
     fi
 
-    yum_install "texinfo"
+    if confirm "I will now build and install gcc from source."; then
+        sudo yum -y install texinfo
 
-    git clone --depth 1 --branch binutils-2_44 \
-        "https://sourceware.org/git/binutils-gdb.git" "$workdir/binutils"
-    cd "$workdir/binutils"
-    mkdir build && cd build
-    CC=gcc10-gcc CXX=gcc10-g++ ../configure --prefix="$HOME/.local" \
-        --disable-gprofng --disable-gdb --disable-gdbserver
-    make -j
-    make install
-    cd "$curdir"
+        git clone --depth 1 --branch binutils-2_44 \
+            "https://sourceware.org/git/binutils-gdb.git" "$workdir/binutils"
+        cd "$workdir/binutils"
+        mkdir build && cd build
+        CC=gcc10-gcc CXX=gcc10-g++ ../configure --prefix="$HOME/.local" \
+            --disable-gprofng --disable-gdb --disable-gdbserver
+        make -j
+        make install
+        cd "$curdir"
 
-    git clone --depth 1 --branch releases/gcc-15 \
-        "https://gcc.gnu.org/git/gcc.git" "$workdir/gcc"
-    cd "$workdir/gcc"
-    ./contrib/download_prerequisites
-    mkdir build && cd build
-    CC=gcc10-gcc CXX=gcc10-g++ ../configure --prefix="$HOME/.local" \
-        --enable-languages=c,c++ --disable-multilib --disable-bootstrap
-    make -j
-    make install
-    cd "$curdir"
+        git clone --depth 1 --branch releases/gcc-15 \
+            "https://gcc.gnu.org/git/gcc.git" "$workdir/gcc"
+        cd "$workdir/gcc"
+        ./contrib/download_prerequisites
+        mkdir build && cd build
+        CC=gcc10-gcc CXX=gcc10-g++ ../configure --prefix="$HOME/.local" \
+            --enable-languages=c,c++ --disable-multilib --disable-bootstrap
+        make -j
+        make install
+        cd "$curdir"
+    fi
 }
 
 sync_node() {
@@ -147,7 +140,7 @@ sync_node() {
         # shellcheck source=/dev/null
         [ -s "$NVM_DIR/nvm.sh" ] && \. "$NVM_DIR/nvm.sh"
 
-        if [ "$os_pkg_manager" = "apt" ]; then
+        if [ "$os" = "apt" ]; then
             nvm install 24
             nvm use 24
             nvm alias default 24
@@ -157,17 +150,6 @@ sync_node() {
         nvm install 16
         nvm use 16
         nvm alias default 16
-        #
-        #if ! [ -x "$(command -v "node")" ]; then
-        #    git clone --depth 1 --branch "v24.14.0" \
-        #        "https://github.com/nodejs/node.git" "$workdir/node"
-        #    cd "$workdir/node"
-        #    CC="$HOME/.local/bin/gcc" CXX="$HOME/.local/bin/g++" \
-        #        ./configure --prefix="$HOME/.nvm/versions/node/v24.10.0"
-        #    make -j
-        #    make install
-        #    cd "$curdir"
-        #fi
     fi
 }
 
@@ -177,64 +159,82 @@ sync_rust() {
     fi
 }
 
-alacritty_install() {
-    if [ -x "$(command -v "alacritty")" ]; then
-        return
+sync_alacritty() {
+    if confirm "I will now replace your alacritty configuration."; then
+        if [ "$os" = "apt" ]; then
+            sudo apt -y install alacritty
+        fi
+        rm -rf "$HOME/.config/alacritty"
+        curl -sSLo "$HOME/.config/alacritty/alacritty.toml" --create-dirs \
+            "$dotfiles_url/home/.config/alacritty/alacritty.toml"
     fi
-    apt_install "cmake"
-    apt_install "g++"
-    apt_install "pkg-config"
-    apt_install "libfontconfig1-dev"
-    apt_install "libxcb-xfixes0-dev"
-    apt_install "libxkbcommon-dev"
-    apt_install "python3"
-    apt_install "libegl1-mesa-dev"
-    git clone --depth 1 "https://github.com/alacritty/alacritty.git" "$workdir/alacritty"
-    cd "$workdir/alacritty"
-    cargo build --release
-    if ! infocmp alacritty; then
-        sudo tic -xe alacritty,alacritty-direct extra/alacritty.info
-    fi
-    sudo cp target/release/alacritty /usr/local/bin
-    sudo cp extra/logo/alacritty-term.svg /usr/share/pixmaps/Alacritty.svg
-    sudo desktop-file-install extra/linux/Alacritty.desktop
-    sudo update-desktop-database
-    cd "$curdir"
 }
 
 sync_cmake() {
-    if [ "$os_pkg_manager" = "apt" ]; then
-        apt_install "libssl-dev"
-        apt_install "cmake"
+    if [ "$os" = "apt" ]; then
+        sudo apt -y install libssl-dev cmake
         return
     fi
-    if [ -x "$(command -v "$HOME/.local/bin/cmake")" ]; then
+    if [ "$os" != "yum" ]; then
+        return
+    fi
+    if [ -x "$HOME/.local/bin/cmake" ]; then
         return
     fi
 
-    yum_install "perl-core"
+    if confirm "I will now build and install cmake from source."; then
+        sudo yum -y install perl-core
 
-    if ! [ -x "$(command -v "$HOME/.local/bin/openssl")" ]; then
-        git clone --depth 1 "https://github.com/openssl/openssl.git" "$workdir/openssl"
-        cd "$workdir/openssl"
-        ./Configure --prefix="$HOME/.local"
-        make -j CC=gcc10-gcc
+        if ! [ -x "$HOME/.local/bin/openssl" ]; then
+            git clone --depth 1 "https://github.com/openssl/openssl.git" "$workdir/openssl"
+            cd "$workdir/openssl"
+            ./Configure --prefix="$HOME/.local"
+            make -j CC=gcc10-gcc
+            make install
+            cd "$curdir"
+        fi
+
+        git clone --depth 1 "https://github.com/Kitware/CMake.git" "$workdir/cmake"
+        cd "$workdir/cmake"
+        ./bootstrap --prefix="$HOME/.local" -- -DOPENSSL_ROOT_DIR="$HOME/.local"
+        make -j
         make install
         cd "$curdir"
     fi
+}
 
-    git clone --depth 1 "https://github.com/Kitware/CMake.git" "$workdir/cmake"
-    cd "$workdir/cmake"
-    ./bootstrap --prefix="$HOME/.local" -- -DOPENSSL_ROOT_DIR="$HOME/.local"
-    make -j
-    make install
-    cd "$curdir"
+sync_ripgrep() {
+    if [ "$os" = "macos" ]; then
+        return
+    fi
+    cargo install ripgrep
+}
+
+sync_tree_sitter() {
+    if [ "$os" = "apt" ]; then
+        sudo apt -y install libgcc-14-dev
+        BINDGEN_EXTRA_CLANG_ARGS="-I/usr/lib/gcc/x86_64-linux-gnu/14/include" \
+            cargo install --locked tree-sitter-cli
+    elif [ "$os" = "yum" ]; then
+        BINDGEN_EXTRA_CLANG_ARGS="-I$HOME/.local/lib/gcc/x86_64-pc-linux-gnu/15.2.1/include" \
+            cargo install --locked tree-sitter-cli
+    fi
+}
+
+sync_nvim_config() {
+    if confirm "I will now replace your vim configuration."; then
+        rm -rf "$HOME/.config/nvim"
+        curl -sSLo "$HOME/.config/nvim/init.lua" --create-dirs \
+            "$dotfiles_url/home/.config/nvim/init.lua"
+    fi
 }
 
 sync_vim() {
-    if [ "$os_pkg_manager" = "apt" ]; then
+    if [ "$os" = "apt" ]; then
         sudo apt-add-repository -y ppa:neovim-ppa/unstable
-        apt_install "neovim"
+        sudo apt -y install neovim clangd
+    elif [ "$os" = "yum" ]; then
+        sudo yum -y install clang-tools-extra
     fi
 
     if ! [ -x "$(command -v "nvim")" ]; then
@@ -248,84 +248,80 @@ sync_vim() {
 
     require_command "nvim"
 
-    apt_install "clangd"
-    yum_install "clang-tools-extra"
-
-    cargo install ripgrep
-
-    if [ "$os_pkg_manager" = "apt" ]; then
-        sudo apt install libgcc-14-dev
-        BINDGEN_EXTRA_CLANG_ARGS="-I/usr/lib/gcc/x86_64-linux-gnu/14/include" \
-            cargo install --locked tree-sitter-cli
-    else
-        BINDGEN_EXTRA_CLANG_ARGS="-I$HOME/.local/lib/gcc/x86_64-pc-linux-gnu/15.2.1/include" \
-            cargo install --locked tree-sitter-cli
-    fi
-
-    if confirm "I will now replace your vim configuration."; then
-        rm -rf "$HOME/.config/nvim"
-        curl -sSLo "$HOME/.config/nvim/init.lua" --create-dirs \
-            "$dotfiles_url/home/.config/nvim/init.lua"
-    fi
+    sync_ripgrep
+    sync_tree_sitter
+    sync_nvim_config
 }
 
-sync_gui() {
-    if [ -z "$XDG_SESSION_TYPE" ]; then
-        echo "I don't think you have a gui..."
-        echo "That's it! Everything is up to date!"
+sync_themes() {
+    if [ "$os" != "apt" ]; then
         return
     fi
-
-    if confirm "All the basic stuff is done. I can now setup the gui."; then
+    if confirm "I will now install papirus icon themes."; then
         sudo add-apt-repository -y ppa:papirus/papirus
         sudo apt update
-
-        apt_install "papirus-icon-theme"
-        apt_install "neovim"
+        sudo apt -y install papirus-icon-theme
 
         papirus-folders -t Papirus -C nordic -u
         papirus-folders -t Papirus-Dark -C nordic -u
+    fi
+}
 
-        mkdir -p "$HOME/.local/share/fonts"
+sync_fonts() {
+    if confirm "I will now install custom fonts."; then
+        if [ "$os" = "macos" ]; then
+            fonts_dir="$HOME/Library/Fonts"
+        else
+            fonts_dir="$HOME/.local/share/fonts"
+        fi
+        mkdir -p "$fonts_dir"
+
         curl -sSLo "$workdir/PerfectDOSVGA437Win.tar.xz" --create-dirs \
             "$dotfiles_url/home/.local/share/fonts/PerfectDOSVGA437Win.tar.xz"
-        tar xf "$workdir/PerfectDOSVGA437Win.tar.xz" -C "$HOME/.local/share/fonts"
+        tar xf "$workdir/PerfectDOSVGA437Win.tar.xz" -C "$fonts_dir"
 
         curl -sSLo "$workdir/AcPlus_IBM_BIOS.tar.xz" --create-dirs \
             "$dotfiles_url/home/.local/share/fonts/AcPlus_IBM_BIOS.tar.xz"
-        tar xf "$workdir/AcPlus_IBM_BIOS.tar.xz" -C "$HOME/.local/share/fonts"
+        tar xf "$workdir/AcPlus_IBM_BIOS.tar.xz" -C "$fonts_dir"
 
-        fc-cache -f
+        if [ "$os" != "macos" ]; then
+            fc-cache -f
+        fi
+    fi
+}
 
-        alacritty_install
-        rm -rf "$HOME/.config/alacritty"
-        curl -sSLo "$HOME/.config/alacritty/alacritty.toml" --create-dirs \
-            "$dotfiles_url/home/.config/alacritty/alacritty.toml"
+sync_apple_keyboard() {
+    if [ "$os" != "apt" ]; then
+        return
     fi
     if confirm "If you're using an apple keyboard driver, I can configure it to act right on linux"; then
         echo 2 | sudo tee /sys/module/hid_apple/parameters/fnmode
         echo "options hid_apple fnmode=2" | sudo tee -a /etc/modprobe.d/hid_apple.conf
         sudo update-initramfs -u -k all
     fi
-    echo "That's it! You should log out and log back in."
 }
 
 sync_bin() {
-    set -- "ssh-tunnel" "modplay" "xsig"
-    for f in "$@"; do
-        curl -sSLo "$HOME/bin/$f" --create-dirs "$dotfiles_url/home/bin/$f"
-        chmod +x "$HOME/bin/$f"
-    done
+    if confirm "I will now replace your ~/bin scripts."; then
+        set -- "ssh-tunnel" "modplay" "xsig" "kiro-acp-proxy"
+        for f in "$@"; do
+            curl -sSLo "$HOME/bin/$f" --create-dirs "$dotfiles_url/home/bin/$f"
+            chmod +x "$HOME/bin/$f"
+        done
+    fi
 }
 
 sync_gdb() {
-    rm -f "$HOME/.gdbinit"
-    curl -sSLo "$HOME/.gdbinit" "$dotfiles_url/home/.gdbinit"
+    if confirm "I will now replace your gdb configuration."; then
+        rm -f "$HOME/.gdbinit"
+        curl -sSLo "$HOME/.gdbinit" "$dotfiles_url/home/.gdbinit"
+    fi
 }
 
 require_command "curl"
 sync_git
 sync_shell
+sync_tmux
 sync_rust
 sync_bin
 sync_gcc
@@ -333,4 +329,7 @@ sync_cmake
 sync_node
 sync_vim
 sync_gdb
-sync_gui
+sync_alacritty
+sync_themes
+sync_fonts
+sync_apple_keyboard
